Completed on Name generation and GET methods

+++ Services/NameService/NameService.cs
@@ -20,7 +20,7 @@ namespace funAPI.Services.NameService
             _mapper = mapper;
         }
 
-        public async Task<ServiceResponse<List<GetNameDTO>>> BookAName(AddNameDTO newName)
+        public async Task<ServiceResponse<List<GetNameDTO>>> AddAName(AddNameDTO newName)
         {
             var serviceResponse = new ServiceResponse<List<GetNameDTO>>();
             Names name = (_mapper.Map<Names>(newName));
@@ -30,45 +30,74 @@ namespace funAPI.Services.NameService
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetNameDTO>>> DeleteAName(int id)
+
+        public async Task<ServiceResponse<List<GetNameDTO>>> DeleteAName(string role, int id)
         {
             var serviceResponse = new ServiceResponse<List<GetNameDTO>>();
-            try
+            if (role.ToUpper() == "ADMIN")
             {
-                Names name = await _context.Names.FirstAsync(n => n.Id == id);
-                _context.Names.Remove(name);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Names.Select(c => _mapper.Map<GetNameDTO>(c)).ToList();
-            }
-            catch (Exception ex)
-            {
-                serviceResponse.Success = false;
-                serviceResponse.Message = ex.Message;
+                try
+                {
+                    Names name = await _context.Names.FirstAsync(n => n.Id == id);
+                    _context.Names.Remove(name);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.Names.Select(c => _mapper.Map<GetNameDTO>(c)).ToList();
+                }
+                catch (Exception ex)
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = ex.Message;
+                }
             }
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetNameDTO>>> GetBookedList(string role)
+        public async Task<ServiceResponse<List<GetNameDTO>>> GenerateAName()
         {
             var serviceResponse = new ServiceResponse<List<GetNameDTO>>();
+            string generatedName = NameGenerator.RandomString(10);
+            Names name = new Names() { Name = generatedName };
+            _context.Names.Add(name);
+            await _context.SaveChangesAsync();
+            serviceResponse.Data = await _context.Names
+            .Where(n => n.DateGenerated.Date == DateTime.UtcNow.Date)
+            .Select(n => _mapper.Map<GetNameDTO>(n))
+            .ToListAsync();
+            return serviceResponse;
+        }
+
+        public async Task<ServiceResponse<List<GetNameDTO>>> GetAll()
+        {
+            var serviceResponse = new ServiceResponse<List<GetNameDTO>>();
+            var dbNames = await _context.Names.ToListAsync();
+            serviceResponse.Data = dbNames.Select(c => _mapper.Map<GetNameDTO>(c)).ToList();
+            return serviceResponse;
+        }
+
+        public async Task<ServiceResponse<List<GetBookedNamesDTO>>> GetBookedList(string role)
+        {
+            var serviceResponse = new ServiceResponse<List<GetBookedNamesDTO>>();
 
             if (role.ToUpper() == "ADMIN")
             {
                 var dbNames = await _context.Names.Where(x => x.IsBooked == true).ToListAsync();
-                serviceResponse.Data = dbNames.Select(c => _mapper.Map<GetNameDTO>(c)).ToList();
+                serviceResponse.Data = dbNames.Select(c => _mapper.Map<GetBookedNamesDTO>(c)).ToList();
                 return serviceResponse;
             }
             else
             {
                 serviceResponse.Data = null;
+                serviceResponse.Message = "Unfortunately you are restricted to access this information";
                 return serviceResponse;
             }
         }
 
-        public async Task<ServiceResponse<List<GetNameDTO>>> GetList()
+        public async Task<ServiceResponse<List<GetNameDTO>>> GetListForToday()
         {
             var serviceResponse = new ServiceResponse<List<GetNameDTO>>();
-            var dbNames = await _context.Names.ToListAsync();
+            var dbNames = await _context.Names
+            .Where(x => x.DateGenerated.Date == DateTime.UtcNow.Date)
+            .ToListAsync();
             serviceResponse.Data = dbNames.Select(c => _mapper.Map<GetNameDTO>(c)).ToList();
             return serviceResponse;
         }
